+++ src/index.ts
@@ -1,2 +1,16 @@
+// 🌐 Entrypoint global du module todo-usecase-default
+// Ce fichier est généré automatiquement. Ne pas modifier à la main.
+
+// 📦 common
+export * from "./common/validation.js";
+
+// 📦 todo-creation
 export * from "./todo-creation/usecases/create-todo.interactor.js";
 export * from "./todo-creation/validations/create-todo.validation.js";
+
+// 📦 todo-modification
+export * from "./todo-modification/usecases/mark-todo-as-completed.interactor.js";
+export * from "./todo-modification/usecases/mark-todo-as-uncompleted.interactor.js";
+
+// 📦 todo-retrieval
+export * from "./todo-retrieval/usecases/get-all-todo.interactor.js";