+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # todo-usecase-default
 
+## 1.4.0
+
+### Minor Changes
+
+- add an index.sh to automate file export
+
+### Patch Changes
+
+- use input to log timestamp
+
+## 1.3.0
+
+### Minor Changes
+
+- new feature : - get all todo - mark todo as completed - mark todo as uncompleted
+
 ## 1.1.0
 
 ### Minor Changes